+++ app/controllers/recipes_controller.rb
@@ -1,4 +1,6 @@
 class RecipesController < ApplicationController
+  before_action :ensure_current_user, only: [:show, :edit, :update, :destroy]
+  
   def new
     @recipe = Recipe.new
     @recipe_categories = RecipeCategory.all
@@ -6,7 +8,6 @@ class RecipesController < ApplicationController
 
   def confirm
     @recipe_categories = RecipeCategory.all
-    # random_num = rand(1..4)
     results = []
     if params[:recipe][:select_recipe_category_id] == "0"
       category_id = params[:recipe][:recipe_category_id]
@@ -14,14 +15,11 @@ class RecipesController < ApplicationController
       category_id = @recipe_categories.offset( rand(@recipe_categories.count) ).first.category_id
     end
     if category_id.blank?
-      # flash.now[:notice] = "カテゴリーを選択するか、完全におまかせを選択してください。"
       redirect_to new_recipe_path, notice: "カテゴリーを選択するか、完全におまかせを選択してください。"
     else
       recipe_category = RecipeCategory.find_by(category_id: category_id)
       RakutenWebService::Recipe.ranking(category_id).each {|n| results << n}
       result = results.sample
-      # result = results.select {|n| n.rank == random_num}
-      # result = results.first
       result_hash = {
         user_id: current_user.id,
         recipe_category_id: recipe_category.id,
@@ -35,12 +33,6 @@ class RecipesController < ApplicationController
       @recipe_image = result["foodImageUrl"]
       @recipe = Recipe.new(result_hash)
     end
-    
-    # if @recipe.recipe_category_id.blank?
-    #   flash.now[:notice] = "カテゴリーを選択するか、完全におまかせを選択してください。"
-    #   render :new
-    # end
-    
   end
   
   def create
@@ -63,21 +55,29 @@ class RecipesController < ApplicationController
   end
 
   def index
-    @recipes = Recipe.page(params[:page])
+    @recipe_categories = RecipeCategory.all
+    if params[:category_id]
+      @recipe_category = RecipeCategory.find(params[:category_id])
+      @category_recipe_all = @recipe_category.recipes.where(user_id: current_user.id)
+      @recipes = @recipe_category.recipes.where(user_id: current_user.id).page(params[:page])
+    else
+      @recipe_all = current_user.recipes.all
+      @recipes = current_user.recipes.page(params[:page])
+    end
   end
 
   def show
-    @recipe = Recipe.find(params[:id])
+    @recipe = current_user.recipes.find(params[:id])
   end
 
   def edit
-    @recipe = Recipe.find(params[:id])
+    @recipe = current_user.recipes.find(params[:id])
   end
   
   def update
-    @recipe = Recipe.find(params[:id])
+    @recipe = current_user.recipes.find(params[:id])
     if @recipe.update(recipe_memo_params)
-      flash.now[:notice] = "レシピのレビューを変更しました"
+      flash[:notice] = "レシピのレビューを変更しました"
       redirect_to recipe_path(@recipe)
     else
       render "edit"
@@ -85,7 +85,7 @@ class RecipesController < ApplicationController
   end
   
   def destroy
-    @recipe = Recipe.find(params[:id])
+    @recipe = current_user.recipes.find(params[:id])
     @recipe.destroy
     redirect_to recipes_path
   end
@@ -98,6 +98,14 @@ class RecipesController < ApplicationController
   
   def recipe_memo_params
     params.require(:recipe).permit(:memo)
+  end
+  
+  def ensure_current_user
+    recipe = Recipe.find(params[:id])
+    unless current_user.id == recipe.user_id
+      flash[:alert] = "他のユーザーのレシピにアクセスする権限がありません。"
+      redirect_to recipes_path
+    end
   end
   
 end
