+++ app/javascript/packs/application.js
@@ -17,15 +17,4 @@ Rails.start()
 Turbolinks.start()
 ActiveStorage.start()
 
-$(document).on('change', '#large_area_results', function() {
-  $.ajax({
-    type: 'GET',
-    url: '/restaurants/middle_area_select',
-    data: {
-      restaurant_large_area: $(this).val()
-    },
-    success: function(data) {
-      $('#middle_area_results').html(data);
-    }
-  });
-});
+import 'packs/area_select';
