+++ app/javascript/packs/area_select.js
@@ -1,0 +1,12 @@
+$(document).on('change', '#large_area_results', function() {
+  $.ajax({
+    type: 'GET',
+    url: '/restaurants/middle_area_select',
+    data: {
+      restaurant_large_area: $(this).val()
+    },
+    success: function(data) {
+      $('#middle_area_results').html(data);
+    }
+  });
+});
