+++ app/views/layouts/application.html.erb
@@ -18,7 +18,16 @@
     <div class = "container px-5 px-sm-0">
       <div class = "row">
         <div class = "col-md-8 mt-2">
-          <%= flash[:notice] %>
+          <% if flash[:notice] %>
+            <div class="alert alert-success">
+              <%= flash[:notice] %>
+            </div>
+          <% end %>
+          <% if flash[:alert] %>
+            <div class="alert alert-danger">
+              <%= flash[:alert] %>
+            </div>
+          <% end %>
         </div>
       </div>
     </div>
