+++ app/views/recipes/index.html.erb
@@ -1,2 +1,42 @@
-<h1>Recipes#index</h1>
-<p>Find me in app/views/recipes/index.html.erb</p>
+<div class="container">
+  <div class="row">
+    <% if params[:category_id] %>
+      <h1 class="col-lg-8 text-center my-5 mx-auto"><%= @recipe_category.name %>一覧(全<%= @category_recipe_all.count %>件)</h5></h1>
+      <%= link_to "レシピ一覧に戻る", recipes_path %>
+    <% else %>
+      <h1 class="col-lg-8 text-center my-5 mx-auto">作ったレシピ一覧(全<%= @recipe_all.count %>件)</h5></h1>
+    <% end %>
+  </div>
+  <div class="row">
+    <div class="col-12">
+      <div class="text-right">
+        <%= form_with model: @recipe, url: recipes_path, method: :get do |f| %>
+          <%= f.collection_select :category_id, @recipe_categories, :id, :name, include_blank: "カテゴリー検索"  %>
+          <%= button_tag type: "submit", class:"btn btn-default" do %>
+            <i class="fa-solid fa-magnifying-glass"></i>
+          <% end %>
+        <% end %>
+      </div>
+    </div>
+  </div>
+  <div class="row">
+    <div class="col-12">
+      <table class='table'> 
+        <tbody>
+          <% @recipes.each do |recipe| %>
+            <tr>
+              <td><%= image_tag recipe.get_recipe_image(500, 500), size: "100x100", class: "rounded" %></td>
+              <td><%= link_to recipe.name, recipe_path(recipe.id) %></td>
+              <td><%= recipe.recipe_category.name %></td>
+              <td></td>
+              <td></td>
+            </tr>
+          <% end %>
+        </tbody>
+      </table>
+      <div class="d-flex justify-content-center mt-3">
+        <%= paginate @recipes %>
+      </div>
+    </div>
+  </div>
+</div>
