+++ config/initializers/kaminari_config.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 Kaminari.configure do |config|
-  # config.default_per_page = 25
+  config.default_per_page = 5
   # config.max_per_page = nil
   # config.window = 4
   # config.outer_window = 0
